wpref/requirements.txt
Django
django-schema-viewer
djangorestframework
django-cors-headers
djangorestframework-simplejwt
drf-spectacular
django-extensions
pillow
django-filter
requests
drf-nested-routers
coverage
django-parler
django-environ
celery[redis]
